% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeGivenYear.R
\name{mergeGivenYear}
\alias{mergeGivenYear}
\title{Create .rda "master" file from .sas7bdat files}
\usage{
mergeGivenYear(dir = getwd(), desired.files = c())
}
\arguments{
\item{dir}{the target directory for the output}

\item{desired.files}{a vector of .sas7bdat file names to merge together}
}
\value{
A data frame of merged .sas7bdat files
}
\description{
Create .rda "master" file from .sas7bdat files
}
