% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predROC.R
\name{predROC}
\alias{predROC}
\title{Produce a data frame of ROC curve values based on a svyglm object and a test set}
\usage{
predROC(glm.obj, newData, plot = TRUE)
}
\arguments{
\item{glm.obj}{a survey::svyglm object whose response is a binary numeric variable taking the values 0, 1 or NA}

\item{newData}{a data frame whose variables are compatible with the svyglm object, for use in prediction}
}
\value{
A data frame of ROC curve values based on a svyglm object and a test set
}
\description{
Produce a data frame of ROC curve values based on a svyglm object and a test set
}
\examples{
data(list = c("GES2013.drivers.design", "GES2013.drivers"))
options(survey.lonely.psu = "adjust")
model <- survey::svyglm(DROWSY ~ HEAVY_TRUCK + INT_HWY + SEX_IM + SPEEDREL,
                            family = quasibinomial(link=logit),
                            design = GES2013.drivers.design)
predROC(model,GES2013.drivers)
}
