% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run shiny app for exploring correlations in GES data sets}
\usage{
runExample()
}
\description{
Run shiny app for exploring correlations in GES data sets
}
\examples{
runExample()
}
