% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_mosaic_splom.R
\name{svy_mosaic_splom}
\alias{svy_mosaic_splom}
\title{Create a grid of 2x2 survey-weighted mosaic plots}
\usage{
svy_mosaic_splom(data, factors = c(), weight = "WEIGHT", svydesignObj,
  mar_par = rep(0.5, 4))
}
\arguments{
\item{data}{a data frame}

\item{factors}{a vector of character strings corresponding to factor-type variables from the data frame}

\item{weight}{a character string indicating the name of the weight variable in the data frame.}

\item{svydesignObj}{a survey::svydesign object, used to create bivariate ssurvey::svyglm models and color the corresponding mosaics red/blue if there is a statistically significant correlation between the two variables}

\item{mar_par}{a numeric vector of length 4 controlling the margins of the plots}
}
\value{
A grid of 2x2 mosaics, one for each pair of the input factors
}
\description{
Create a grid of 2x2 survey-weighted mosaic plots
}
\examples{
data(list = c("GES2013.drivers.design", "GES2013.drivers"))
factors <- c("DROWSY", "HEAVY_TRUCK",  "INT_HWY",  "SEX_IM", "SPEEDREL")
svy_mosaic_splom(data = GES2013.drivers, factors = factors, weight = "WEIGHT", svydesignObj = GES2013.drivers.design)
}
